[CK-0] add ADR-008 3-layer architecture and align auth code: UserManager, JWTManager, OAuthManager, AuthManager (root)
